requirements.txt
langchain
pinecone-client
openai
streamlit
beautifulsoup4